Added CoinFilter and Quarter Test

+++ src/main/java/com/natevory/kata/vending/Coin.java
@@ -2,27 +2,27 @@ package com.natevory.kata.vending;
 
 
 public class Coin {
-	private int weight, size;
+	private double weight, size;
 	
-	public Coin(int weight, int size){
+	public Coin(double weight, double size){
 		setWeight(weight);
 		setSize(size);
 	}
-	public int getWeight(){
+	public double getWeight(){
 		return weight;
 	}
-	public int getSize(){
+	public double getSize(){
 		return size;
 	}
 	
 	
-	private void setWeight(int weight){
+	private void setWeight(double weight){
 		if(weight<=0)
 			throw new IllegalArgumentException("Coin weight must be greater than 0");
 		this.weight=weight;
 	}
 	
-	private void setSize(int size){
+	private void setSize(double size){
 		if(size<=0)
 			throw new IllegalArgumentException("Coin size must be greater than 0");
 		this.size=size;
